// config/env.enum.ts
export enum EnvKeysEnum {
  NODE_ENV = 'NODE_ENV',
  SERVER_HTTP_PORT = 'SERVER_HTTP_PORT',
  SERVER_HTTP_HOST = 'SERVER_HTTP_HOST',
  DB_DIALECT = 'DB_DIALECT',
  DB_PORT = 'DB_PORT',
  DB_DATABASE = 'DB_DATABASE',
  DB_HOST = 'DB_HOST',
  DB_USERNAME = 'DB_USERNAME',
  DB_PASSWORD = 'DB_PASSWORD',
  DB_AUTO_LOAD_MODELS = 'DB_AUTO_LOAD_MODELS',
  DB_SYNC = 'DB_SYNC',
  DB_FORCE = 'DB_FORCE',
  DB_POOL_MIN = 'DB_POOL_MIN',
  DB_POOL_MAX = 'DB_POOL_MAX',
  DB_LOGGING = 'DB_LOGGING',
  DB_UNDERSCORED = 'DB_UNDERSCORED',
  DB_SSL = 'DB_SSL',
  KAFKA_CONSUMER_BROKERS = 'KAFKA_CONSUMER_BROKERS',
  KAFKA_PRODUCER_BROKERS = 'KAFKA_PRODUCER_BROKERS',
  KAFKA_CLIENT_ID = 'KAFKA_CLIENT_ID',
  KAFKA_ALLOW_AUTO_TOPIC_CREATION = 'KAFKA_ALLOW_AUTO_TOPIC_CREATION',
  KAFKA_AUTO_COMMIT = 'KAFKA_AUTO_COMMIT',
  KAFKA_FROM_BEGINING = 'KAFKA_FROM_BEGINING',
  KAFKA_AUDIT_LOG_TOPIC = 'KAFKA_AUDIT_LOG_TOPIC',
  IS_SWAGGER_ENABLED = 'IS_SWAGGER_ENABLED',
  GRAVITEE_ENDPOINT = 'GRAVITEE_ENDPOINT',
  CONSUMER_IDENTITY_SERVICE_URL = 'CONSUMER_IDENTITY_SERVICE_URL',
  OTEL_EXPORTER_OTLP_ENDPOINT = 'OTEL_EXPORTER_OTLP_ENDPOINT',
  SERVICE_NAME = 'SERVICE_NAME',
}